diet-coaching-chatbot/requirements.txt
freezegun~=1.1.0
requests~=2.23.0
PyYAML~=6.0
rasa~=2.7.0
rasa_sdk~=2.7.0
numpy~=1.16
scikit-learn~=0.24.2
matplotlib~=3.3.4
pandas~=1.3.5
mysql-connector-python~=8.0.27
Jinja2~=3.0.3
packaging~=20.0
pathos~=0.2.8
fuzzywuzzy~=0.18.0
inflect~=5.3.0
myfitnesspal~=2.1.0
statsmodels~=0.13.1
pyngrok~=7.1.2
matplotlib-label-lines~=0.5.1
protobuf~=3.19.0
datefinder
suffix-trees
time_machine
python-dotenv~=0.21.1
urllib3
jsonpickle~=2.0.0
tqdm~=4.59.0
joblib~=1.0.1
google-api-python-client~=2.118.0
oauth2client~=4.1.3
gspread~=6.0.2
python-Levenshtein~=0.25.1
git+https://github.com/huggingface/transformers
torch
torchvision
torchaudio
accelerate
pymysql